refactor: update build config for @primer/mcp to rolldown (#8042)
