+++ packages/doc-gen/package.json
@@ -38,6 +38,7 @@
   "devDependencies": {
     "rimraf": "^6.0.1",
     "rolldown": "^1.1.2",
-    "rolldown-plugin-dts": "^0.26.0"
+    "rolldown-plugin-dts": "^0.26.0",
+    "typescript": "^6.0.3"
   }
 }
